Create view model and repository for home fragment and update recycler adapter

+++ app/src/main/java/com/gomezdevlopment/chessnotationapp/view/home_screen/HomeFragment.kt
@@ -1,6 +1,5 @@
 package com.gomezdevlopment.chessnotationapp.view.home_screen
 
-import android.app.AlertDialog
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
@@ -8,27 +7,29 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.core.content.ContextCompat
+import android.widget.Toast
 import androidx.core.content.FileProvider
-import androidx.fragment.app.FragmentTransaction
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.Navigation
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.gomezdevlopment.chessnotationapp.R
-import com.gomezdevlopment.chessnotationapp.view.MainActivity.Companion.blackAnnotations
-import com.gomezdevlopment.chessnotationapp.view.MainActivity.Companion.whiteAnnotations
-import com.gomezdevlopment.chessnotationapp.view.move_input_feature.MoveAdapter
+import com.gomezdevlopment.chessnotationapp.adapter.MoveAdapter
 import com.gomezdevlopment.chessnotationapp.databinding.FragmentHomeBinding
-import com.gomezdevlopment.chessnotationapp.view.MainActivity
-import com.gomezdevlopment.chessnotationapp.view.move_input_feature.AddNotationFragment
+import com.gomezdevlopment.chessnotationapp.view_model.HomeViewModel
 import java.io.File
 
 class HomeFragment : Fragment() {
 
     private lateinit var binding: FragmentHomeBinding
+    private lateinit var homeViewModel: HomeViewModel
+    private lateinit var whiteMovesAdapter: MoveAdapter
+    private lateinit var blackMovesAdapter: MoveAdapter
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
+        homeViewModel = ViewModelProvider(this).get(HomeViewModel::class.java)
+        homeViewModel.init()
     }
 
     override fun onCreateView(
@@ -47,70 +48,68 @@ class HomeFragment : Fragment() {
         val blackMovesRecycler = binding.blackMovesRecycler
 
         whiteMovesRecycler.layoutManager = LinearLayoutManager(view.context)
-        val whiteMovesAdapter = MoveAdapter(whiteAnnotations)
+        whiteMovesAdapter = MoveAdapter(arrayListOf())
         whiteMovesRecycler.adapter = whiteMovesAdapter
 
         blackMovesRecycler.layoutManager = LinearLayoutManager(view.context)
-        val blackMovesAdapter = MoveAdapter(blackAnnotations)
+        blackMovesAdapter = MoveAdapter(arrayListOf())
         blackMovesRecycler.adapter = blackMovesAdapter
 
+        homeViewModel.getBlackMoves().observe(viewLifecycleOwner, Observer {
+            blackMovesAdapter.addData(it)
+        })
+
+        homeViewModel.getWhiteMoves().observe(viewLifecycleOwner, Observer {
+            whiteMovesAdapter.addData(it)
+        })
+
         binding.newGameButton.setOnClickListener {
-            createNewGameAlert(view.context)
+            homeViewModel.createNewGameDialog(view.context)
         }
 
         binding.exportButton.setOnClickListener {
-            if (whiteAnnotations.isNotEmpty()) {
+            if (whiteMovesAdapter.itemCount > 0) {
                 exportGameFile(view.context)
             }
         }
 
         binding.floatingActionButton.setOnClickListener {
-            Navigation.findNavController(view).navigate(R.id.action_homeFragment_to_addNotationFragment)
+            Navigation.findNavController(view)
+                .navigate(R.id.action_homeFragment_to_addNotationFragment)
         }
-    }
-
-    private fun createNewGameAlert(context: Context) {
-        val builder = AlertDialog.Builder(context)
-        builder.setTitle("Create a new game?")
-        builder.setMessage("The current game will be deleted.")
-
-        builder.setPositiveButton("Yes") { dialog, which ->
-            binding.blackMovesRecycler.adapter?.notifyItemRangeRemoved(0, blackAnnotations.size)
-            binding.whiteMovesRecycler.adapter?.notifyItemRangeRemoved(0, whiteAnnotations.size)
-            whiteAnnotations.clear()
-            blackAnnotations.clear()
-        }
-
-        builder.setNegativeButton("Cancel") { dialog, which ->
-
-        }
-        builder.show()
     }
 
     private fun exportGameFile(context: Context) {
-        val filename = "chess_game.pgn"
-        val path = activity?.getExternalFilesDir(null)
-        val pgnFile = File(path, filename)
-        pgnFile.delete()
-        pgnFile.createNewFile()
+        if (homeViewModel.getWhiteMoves().value.isNullOrEmpty()) {
+            println("Empty")
+            Toast.makeText(context, "Cannot export game because there are no annotations!", Toast.LENGTH_SHORT).show()
+        } else {
+            if (whiteMovesAdapter.itemCount != blackMovesAdapter.itemCount) {
+                homeViewModel.addBlackMove("")
+            }
+            val filename = "chess_game.pgn"
+            val path = activity?.getExternalFilesDir(null)
+            val pgnFile = File(path, filename)
+            pgnFile.delete()
+            pgnFile.createNewFile()
 
-        if (whiteAnnotations.size != blackAnnotations.size) {
-            blackAnnotations.add("")
-        }
+            val whiteMoves = homeViewModel.getWhiteMoves().value!!
+            val blackMoves = homeViewModel.getBlackMoves().value!!
 
-        for (notation in whiteAnnotations) {
-            val index = whiteAnnotations.indexOf(notation)
-            pgnFile.appendText("${index + 1}. $notation ${blackAnnotations[index]} ")
+            for (notation in whiteMoves) {
+                val index = whiteMoves.indexOf(notation)
+                pgnFile.appendText("${index + 1}. $notation ${blackMoves[index]} ")
+            }
+            val uri = FileProvider.getUriForFile(
+                context,
+                activity?.applicationContext?.packageName.toString() + ".provider",
+                pgnFile
+            )
+            val sendIntent = Intent(Intent.ACTION_SEND)
+            sendIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
+            sendIntent.putExtra(Intent.EXTRA_STREAM, uri)
+            sendIntent.type = "text/pgn"
+            startActivity(Intent.createChooser(sendIntent, "SHARE"))
         }
-        val uri = FileProvider.getUriForFile(
-            context,
-            activity?.applicationContext?.packageName.toString() + ".provider",
-            pgnFile
-        )
-        val sendIntent = Intent(Intent.ACTION_SEND)
-        sendIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
-        sendIntent.putExtra(Intent.EXTRA_STREAM, uri)
-        sendIntent.type = "text/pgn"
-        startActivity(Intent.createChooser(sendIntent, "SHARE"))
     }
 }
